Created LineChart component
